auresamp: fix clang warning "loses integer precision"

+++ modules/auresamp/auresamp.c
@@ -85,7 +85,7 @@ static int sampv_alloc(struct auresamp_st *st, struct auframe *af)
 
 static int rsampv_check_size(struct auresamp_st *st, struct auframe *af)
 {
-	uint32_t ptime;
+	uint64_t ptime;
 	size_t psize;
 
 	ptime = af->sampc * 1000 / af->srate;
